requirements.txt
django_tables2
appdirs==1.4.3
gunicorn==19.7.1
packaging==16.8
pyparsing==2.2.0
six==1.10.0
whitenoise==3.3.0
Django==2.2.4
django-storages==1.5.2
olefile==0.44

python-dateutil==2.6.0
pytz==2016.10
Unidecode==0.4.20
virtualenv==15.1.0
django-crispy-forms
django-smartfields
python3-memcached==1.51
boto==2.46.1
boto3==1.4.4
django-import-export==1.0.0
python-decouple==3.0
django-mptt==0.9.0
django-tinymce
psycopg2-binary
unicode-slugify==0.1.3

model-mommy==1.6.0
python-social-auth
social-auth-app-django==3.4.0
pika==0.9.14
django-paypal==1.0.0
celery==4.3
django-celery-beat==2.0.0
django-celery-results==1.2.1
redis
django-sendgrid-v5==0.8.0
reportlab==3.5.42
xlwt==1.3.0
django-autocomplete-light==3.5.1

